+++ app/files/storage_strategies/storage_strategy.py
