+++ app/ocr_strategies2/marker.py
@@ -1,7 +1,7 @@
 from marker.convert import convert_single_pdf
 from marker.models import load_all_models
 
-from ocr_strategies.ocr_strategy import OCRStrategy
+from extract.ocr_strategies.ocr_strategy import OCRStrategy
 
 class MarkerOCRStrategy(OCRStrategy):
     """Marker OCR Strategy"""
